Move symbol types into enums module
