+++ src/main/kotlin/com/github/derg/transpiler/phases/resolver/Expressions.kt
@@ -1,7 +1,7 @@
 package com.github.derg.transpiler.phases.resolver
 
+import com.github.derg.transpiler.source.*
 import com.github.derg.transpiler.source.ast.*
-import com.github.derg.transpiler.source.lexeme.*
 import com.github.derg.transpiler.source.thir.*
 import com.github.derg.transpiler.utils.*
